buggy server reconciliation stub

+++ core/src/otechniques/objects/Player.java
@@ -1,6 +1,6 @@
 package otechniques.objects;
 
 public class Player{
-	public int x, y;
+	public float x, y;
 }
